+++ Classes/Decorators/base_decorator.rb
