+++ Classes/Decorators/capitalize_decorator.rb
