+++ Classes/Decorators/nameable.rb
