+++ Classes/Decorators/trimmed_decorator.rb
